Bug Fix: Trailing white space causes extra row with no entry
After some research, I determined that the splice method will include the separator if it is at the end (the pesky trailing whitespaces). To remedy this, I revisited the regular expression and adjusted it so that it completely removed any whitespaces at the end of the string. I also discovered a really cool site (https://regex101.com/) that assists you with the truly recondite syntax for regular expressions (I mean, what non-computer derived hooman bean can intuitively understand that mess of letters??).

I also adjusted the parameter in removeAllChildNodes() for clarity since we're removing child nodes from the node, not its parent.