Added feature
Repurposed a snippet of code found from StackOverflow to parse string content from a .txt file. Added an event handler (a block of code that runs when an event is dispatched) called readFile, which takes an event as its parameter. Using the event's target (a reference to the DOM element the event acted/was dispatched on), this function uses FileReader to convert the contents of the uploaded file to a string that is stored in the 'result' property. Included another event handler that replaces the textbox value with the fileReader results once the file has been read successfully (onload).

Originally used addEventListener in the script but later determined that it is cleaner to add it as an attribute to the DOM element (less stacked function calls).